FIX: tweepyshell: ipython shell took original arguments
Previously the ipython shell kidnapped all tweepyshell input
arguments and interpreted as its own. Thus e.g. supplying
a username resulted in it trying to open a file with a name equal
to the username. This is not nice so let's put an end to it by
explicitly passing an empty arguments list.