Return cursors when calling method from DMCursorIterator
This adds a return_cursors kwarg and attribute to APIMethod that can be used to explicitly specify to return cursors with the result.
The existing check for the cursor session parameter has been refactored from JSONParser.parse to APIMethod.execute.
This means that the parser's parse method is called with a return_cursors kwarg.
All parsers in the library have been updated to reflect this change.