Implementation of the Account Activity API
This has been in use as a sub classed module for around a month. This is taking that sub classed module and dropping it into the tweepy module with barely no difference other than importing the required dependencies.

I have not decorated the individual API calls, as properties. I don't think there's any effective change whether they are used in this context or not.

In retrospect I should have been more consistent in the naming of the functions.