Bugfix when checkine None mimetypes
When mimetypes.guess_type encounters a type it is unfamiliar with,
it reuturns the tuple (None, None). In api.py, when we check for an
unrecognized filetype, it is before the assignment
`file_type = file_type[0]`. As a result, this block always passes,
even when an unrecognized filetype is encountered.

This change hoists the file_type assignemnt to the first value of the
tuple a few lines higher to remedy the issue.