Remove max file size checks for chunked uploads
The error messages Twitter's API responds with when the file size is too large, "File size exceeds 15728640 bytes." for GIFs and "File size exceeds 536870912 bytes." for videos, are fairly self-explanatory.
This check was incorrect for videos anyway, which seem to have a size restriction of 512 MiB, not 15 MB.
https://developer.twitter.com/en/docs/twitter-api/v1/media/upload-media/uploading-media/media-best-practices
https://twittercommunity.com/t/conflicting-documented-size-restrictions-for-video-uploading-via-api/148201
Allowing Twitter's API to handle the size restrictions and avoiding hardcoding them also removes the need to update them in the future if they change.