Remove DM event section in docs/expansions_and_fields.rst
The documentation for dm_event.fields is different and there are no endpoints with that parameter that return other types of objects, so this section is unnecessary.

@Harmon758 have plans to improve this documentation as a whole to include documentation from https://developer.twitter.com/en/docs/twitter-api/expansions, so for now, for consistency, this section shouldn't be here.

But if we want some completeness in this doc page before the next big doc improvement (and also for the consistency in Client methods' docstrings), we can put the following code back later just in case people thought the doc is not up-to-date:

```rst
.. _dm_event_fields_parameter:

``dm_event_fields``
---------------

Extra fields to include in the event payload. `id`, `text` and `event_type` are
returned by default.

.. _Direct Message event fields: https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/dm-events
```

Co-Authored-By: Harmon <Harmon.Public@gmail.com>